skip running get_discovery_image task for cdn

+++ automation_tools/utils.py
@@ -100,34 +100,37 @@ def get_discovery_image():
     """ Task for getting unattended foreman-discovery ISO image
     :return: foreman-discovery-image iso under /var/lib/libvirt/images/
     """
-    url = os.environ.get('BASE_URL') + '/Packages/'
-    soup = BeautifulSoup(urlopen(url).read())
-    for link in soup.findAll('a'):
-        if 'foreman-discovery-image' in link.string:
-            discovery_image = link.string
-    try:
-        run("wget -O /tmp/" + discovery_image + " " + url + discovery_image)
-        run('cd /tmp/ ; rpm2cpio ' + discovery_image + '|cpio -idmv')
-        run('cp /tmp/usr/share/foreman-discovery-image/'
-            + discovery_image.split('.el')[0] + '.iso /tmp/')
-        run('cp /tmp/usr/bin/discovery-remaster /tmp/')
-        run('/tmp/discovery-remaster /tmp/' + discovery_image.split('.el')[0]
-            + '.iso "fdi.pxgw=' + os.environ.get('GATEWAY') +
-            ' fdi.pxdns=$(cat /etc/resolv.conf|grep -i "^nameserver"|'
-            'head -n1|cut -d " " -f2) proxy.url=https://'
-            + os.environ.get('IPADDR') +
-            ':9090 proxy.type=proxy fdi.pxfactname1=myfact '
-            'fdi.pxfactvalue1=somevalue fdi.pxauto=1" /var/lib/libvirt/images/'
-            + os.environ.get('DISCOVERY_ISO'))
-        size = run('du -h "/var/lib/libvirt/images/"' +
-                   os.environ.get('DISCOVERY_ISO')
-                   + ' | cut -f1 | tr -d [:alpha:]')
-        if int(size) < 150:
-            raise Exception("Generated ISO size is less than 150M!"
-                            " Check if ISO is corrupted.")
-    finally:
-        run('rm /tmp/foreman-discovery-image* /tmp/discovery-remaster '
-            '/tmp/usr -rvf')
+    if os.environ.get('BASE_URL') is not None:
+        url = os.environ.get('BASE_URL') + '/Packages/'
+        soup = BeautifulSoup(urlopen(url).read())
+        for link in soup.findAll('a'):
+            if 'foreman-discovery-image' in link.string:
+                discovery_image = link.string
+        try:
+            run("wget -O /tmp/" + discovery_image + " " + url + discovery_image)
+            run('cd /tmp/ ; rpm2cpio ' + discovery_image + '|cpio -idmv')
+            run('cp /tmp/usr/share/foreman-discovery-image/'
+                + discovery_image.split('.el')[0] + '.iso /tmp/')
+            run('cp /tmp/usr/bin/discovery-remaster /tmp/')
+            run('/tmp/discovery-remaster /tmp/' + discovery_image.split('.el')[0]
+                + '.iso "fdi.pxgw=' + os.environ.get('GATEWAY') +
+                ' fdi.pxdns=$(cat /etc/resolv.conf|grep -i "^nameserver"|'
+                'head -n1|cut -d " " -f2) proxy.url=https://'
+                + os.environ.get('IPADDR') +
+                ':9090 proxy.type=proxy fdi.pxfactname1=myfact '
+                'fdi.pxfactvalue1=somevalue fdi.pxauto=1" /var/lib/libvirt/images/'
+                + os.environ.get('DISCOVERY_ISO'))
+            size = run('du -h "/var/lib/libvirt/images/"' +
+                       os.environ.get('DISCOVERY_ISO')
+                       + ' | cut -f1 | tr -d [:alpha:]')
+            if int(size) < 150:
+                raise Exception("Generated ISO size is less than 150M!"
+                                " Check if ISO is corrupted.")
+        finally:
+            run('rm /tmp/foreman-discovery-image* /tmp/discovery-remaster '
+                '/tmp/usr -rvf')
+    else:
+        print("Skipping...URL for discovery image not found!")
 
 
 def get_packages_name(html):
